% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_eeg.R
\docType{data}
\name{con_eeg}
\alias{con_eeg}
\title{EEG coordinates and potentials for control subjects}
\format{
\subsection{\code{con_eeg}}{

A data frame with 57 rows and 8 columns:
\describe{
\item{x}{projected x axis}
\item{y}{projected y axis}
...
}
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/121/eeg+database}
}
\usage{
con_eeg
}
\description{
A subset of data from EEG experiment to test genetic predisposition to alcoholism
}
\keyword{datasets}
