% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spt_data.R
\name{generate_spt_data}
\alias{generate_spt_data}
\title{Generates simulated space-time data}
\usage{
generate_spt_data(
  Ns = NULL,
  Nt = NULL,
  pattern.type = c("1", "2", "3", "4"),
  tau = 1,
  gradients = TRUE,
  derived.geom = TRUE,
  seed = NULL,
  grid.points = NULL
)
}
\arguments{
\item{Ns}{number of spatial points}

\item{Nt}{number of temporal points}

\item{pattern.type}{four different choices: 1, 2, 3 and 4}

\item{tau}{error std. deviation}

\item{gradients}{logical computes only gradients if TRUE}

\item{derived.geom}{logical computes differential geometric quantities like divergence and Laplacian}

\item{seed}{sets seed}

\item{grid.points}{grid points for computing true values}
}
\description{
Generates simulated space-time data
}
\keyword{generate_spt_data}
