% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsed-mh-spt.R
\name{hlmBayes_mh.spt}
\alias{hlmBayes_mh.spt}
\title{A collapsed spatiotemporal hierarchical Bayes Markov Chain Monte Carlo (MCMC) sampler}
\usage{
hlmBayes_mh.spt(
  y = NULL,
  coords = NULL,
  t = NULL,
  niter = NULL,
  nburn = NULL,
  report = NULL,
  cov.type = NULL,
  a.sigma = NULL,
  b.sigma = NULL,
  a.tau = NULL,
  b.tau = NULL,
  lower.nu = NULL,
  upper.nu = NULL,
  lower.phis = NULL,
  upper.phis = NULL,
  lower.phit = NULL,
  upper.phit = NULL,
  verbose = TRUE,
  trgt_fn.compute = FALSE,
  digits = 3
)
}
\arguments{
\item{y}{observed response (order \eqn{N} x  \eqn{1}), \eqn{N= N_s\times N_t}}

\item{coords}{spatial coordinates for observed process (order \eqn{N_s} x \eqn{2})}

\item{t}{temporal coordinates for observed process (order \eqn{N_t} x \eqn{1})}

\item{niter}{number of MCMC iterations}

\item{nburn}{number of burn-in samples}

\item{report}{batch length}

\item{cov.type}{covariance type (three available choices: Gaussian, Mat\'ern(\eqn{\nu=1/2}), Mat\'ern(\eqn{\nu=3/2}), Mat\'ern(\eqn{\nu=5/2})}

\item{a.sigma}{shape parameter for inverse-gamma prior on \eqn{\sigma^2}}

\item{b.sigma}{scale parameter for inverse-gamma prior on \eqn{\sigma^2}}

\item{a.tau}{shape parameter for inverse-gamma prior on \eqn{\tau^2}}

\item{b.tau}{scale parameter for inverse-gamma prior on \eqn{\tau^2}}

\item{lower.nu}{lower bound for fractal parameter}

\item{upper.nu}{lower bound for fractal parameter}

\item{lower.phis}{lower bound for uniform prior on \eqn{\phi_s}}

\item{upper.phis}{upper bound for uniform prior on \eqn{\phi_s}}

\item{lower.phit}{lower bound for uniform prior on \eqn{\phi_t}}

\item{upper.phit}{upper bound for uniform prior on \eqn{\phi_t}}

\item{verbose}{logical prints output for batches}

\item{trgt_fn.compute}{compute value of target function}

\item{digits}{number of significant digits}
}
\description{
Fits the univariate Gaussian spatiotemporal regression model: \eqn{Y(s,t)=x(s,t)^T\beta+Z(s,t)+\epsilon}. Parameters not listed are optional.
}
\details{
The collapsed posterior being,

\eqn{U(\phi_s|a_{\phi_s},b_{\phi_s}) \times U(\phi_t|a_{\phi_t},b_{\phi_t}) \times IG(\sigma^2|a_{\sigma},b_{\sigma})\times IG(\tau^2|a_{\tau},b_{\tau}) \times N(\beta|\mu_{\beta},\sigma_{\beta}^2) \times  \mathcal{N}(y(s,t)|x(s,t)^T\beta,\sigma^2 R(\phi_s,\phi_t)+\tau^2I)}
}
\keyword{hlmBayes_mh.spt}
