% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-mh-spt-sepbl.R
\name{hlmBayes_spt}
\alias{hlmBayes_spt}
\title{A spatiotemporal hierarchical Bayes Markov Chain Monte Carlo sampler}
\usage{
hlmBayes_spt(
  coords = NULL,
  t = NULL,
  y = NULL,
  X = NULL,
  z_init = NULL,
  phis_init = NULL,
  lower_phis = NULL,
  upper_phis = NULL,
  phit_init = NULL,
  lower_phit = NULL,
  upper_phit = NULL,
  sigma2_init = NULL,
  shape_sigma = NULL,
  scale_sigma = NULL,
  tau2_init = NULL,
  shape_tau = NULL,
  scale_tau = NULL,
  beta_init = NULL,
  mean_beta = NULL,
  prec_beta = NULL,
  steps_init = NULL,
  stept_init = NULL,
  niter = NULL,
  nburn = NULL,
  report = NULL,
  verbose = FALSE,
  digits = NULL,
  cov.type.s = c("exponential", "gaussian", "matern1", "matern2"),
  cov.type.t = c("exponential", "gaussian", "matern1", "matern2")
)
}
\arguments{
\item{coords}{coordinates for observed process (order \eqn{N_s} x \eqn{2})}

\item{t}{temporal coordinates for observed process (order \eqn{N_t} x \eqn{1})}

\item{y}{observed response (order \eqn{N} x  \eqn{1}), \eqn{N= N_s\times N_t}}

\item{X}{a matrix of covariates (order \eqn{N} x  \eqn{p})}

\item{z_init}{starting values of spatial effects (order \eqn{N} x  \eqn{1})}

\item{phis_init}{starting value for \eqn{\phi_s}}

\item{lower_phis}{lower bound for uniform prior on \eqn{\phi_s}}

\item{upper_phis}{upper bound for uniform prior on \eqn{\phi_s}}

\item{phit_init}{starting value for \eqn{\phi_t}}

\item{lower_phit}{lower bound for uniform prior on \eqn{\phi_t}}

\item{upper_phit}{upper bound for uniform prior on \eqn{\phi_t}}

\item{sigma2_init}{starting value for \eqn{\sigma^2}}

\item{shape_sigma}{shape parameter for inverse-gamma prior on \eqn{\sigma^2}}

\item{scale_sigma}{scale parameter for inverse-gamma prior on \eqn{\sigma^2}}

\item{tau2_init}{starting value for \eqn{\tau^2}}

\item{shape_tau}{shape parameter for inverse-gamma prior on \eqn{\tau^2}}

\item{scale_tau}{scale parameter for inverse-gamma prior on \eqn{\tau^2}}

\item{beta_init}{starting value for \eqn{\beta}}

\item{mean_beta}{mean parameter for normal prior on \eqn{\beta}}

\item{prec_beta}{precision (1/variance) parameter for normal prior on \eqn{\beta}}

\item{steps_init}{tuning parameter for \eqn{\phi_s}}

\item{stept_init}{tuning parameter for \eqn{\phi_t}}

\item{niter}{number of MCMC iterations}

\item{nburn}{number of burn-in samples}

\item{report}{batch length}

\item{verbose}{if true prints output for batches}

\item{digits}{number of significant digits}

\item{cov.type.s}{covariance type (three available choices: Exponential, Gaussian, Mat\'ern(\eqn{\nu=1/2}), Mat\'ern(\eqn{\nu=3/2}), Mat\'ern(\eqn{\nu=5/2})}

\item{cov.type.t}{covariance type (three available choices: Exponential, Gaussian, Mat\'ern(\eqn{\nu=1/2}), Mat\'ern(\eqn{\nu=3/2}), Mat\'ern(\eqn{\nu=5/2}), Polynomial}
}
\description{
Fits a univariate Gaussian spatiotemporal regression model: \eqn{Y(s,t)=x(s,t)^T\beta+Z(s,t)+\epsilon}. Parameters not listed are optional.
}
\details{
The posterior being,

\eqn{U(\phi_s|a_{\phi_s},b_{\phi_s})\times U(\phi_t|a_{\phi_t},b_{\phi_t}) \times IG(\sigma^2|a_{\sigma},b_{\sigma})\times IG(\tau^2|a_{\tau},b_{\tau})\times N(\beta|\mu_{\beta},\sigma_{\beta}^2)\times N_{N_s}(Z|0,\Sigma)\times\prod_{i_s=1}^{N_s} \prod_{i_t=1}^{N_t} N(y(s_{i_s},t_{i_t})|x(s_{i_s},t_{i_t})^T\beta+Z(s_{i_s},t_{i_t}),\tau^2)}
}
\keyword{hlmBayes_spt}
