% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_c.R
\name{partition_curve}
\alias{partition_curve}
\title{Partitions a planar curve in space}
\usage{
partition_curve(curve.list = NULL, length.p = 10)
}
\arguments{
\item{curve.list}{a list of planar curves}

\item{length.p}{length of partition}
}
\description{
Partitions a planar curve in space
}
\keyword{partition_curve}
