% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_ge_plot.R
\name{plot_diff.geo}
\alias{plot_diff.geo}
\title{Plots estimated differential geometric quantities}
\usage{
plot_diff.geo(coords = NULL, diff.qty = NULL, col.vec = NULL)
}
\arguments{
\item{coords}{coordinates of grid points}

\item{diff.qty}{differential geometric quantity}

\item{col.vec}{color vector for significance}
}
\description{
Uses ggplot2 to generate estimated plots for spatotemporal divergence and Laplacian
}
\keyword{plot_diff.geo}
