% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcmc.R
\name{plot_mcmc}
\alias{plot_mcmc}
\title{Function for plotting convergence diagnostics for MCMC}
\usage{
plot_mcmc(
  samples = NULL,
  true = NULL,
  col = "darkgreen",
  acf.lag.max = 100,
  cnames = NULL,
  diagnostic.measures = TRUE
)
}
\arguments{
\item{samples}{posterior MCMC samples}

\item{true}{true values, if available}

\item{col}{color of median lines}

\item{acf.lag.max}{maximum lag for ACF plot}

\item{cnames}{names of parameters}

\item{diagnostic.measures}{logical computes Geweke diagnostic for convergence}
}
\description{
Function for plotting convergence diagnostics for MCMC
}
\keyword{plot_mcmc}
