% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-surf-wm-sig.R
\name{plot_surf_tri_sig}
\alias{plot_surf_tri_sig}
\title{Plots significance at the triangular region level}
\usage{
plot_surf_tri_sig(
  curves_part = NULL,
  tr.points = 10,
  length.p = 10,
  wm.signif = NULL
)
}
\arguments{
\item{curves_part}{partitioned curves}

\item{tr.points}{number of points in triangle (should be fixed at 10 for now)}

\item{length.p}{length of partition (should be fixed at 10 for now)}

\item{wm.signif}{wombling measure estimated at the triangular level (should be a data.frame)}
}
\description{
Plots significance at the triangular region level
}
\keyword{plot.surf.wm.sig}
