% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm2.5_final.noNAsepoct.R
\docType{data}
\name{pm2.5_final.noNAsepoct}
\alias{pm2.5_final.noNAsepoct}
\title{PM2.5 data for Sep 29--Oct 2, 2023 in north Mid-Atlantic and New England}
\format{
\subsection{\code{pm2.5_final.noNAsepoct}}{

A data frame with 154 rows and 8 columns:
\describe{
\item{State}{Name of State}
\item{Site.Name}{Name of Station}
\item{id}{ID of Station}
\item{lat}{Latitude}
\item{long}{Longitude}
...
}
}
}
\source{
\url{https://www.epa.gov/outdoor-air-quality-data/download-daily-data}
}
\usage{
pm2.5_final.noNAsepoct
}
\description{
Air quality data for the Canada wildfire smoke waves for Sep-Oct
}
\keyword{datasets}
