% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_finalCA.R
\docType{data}
\name{precip_finalCA}
\alias{precip_finalCA}
\title{Precipitation in California}
\format{
\subsection{\code{precip_finalCA}}{

A data frame with 1392 rows and 18 columns:
\describe{
\item{state}{Name of State}
\item{name}{Name of Station}
\item{id}{ID of Station}
\item{lat}{Latitude}
\item{long}{Longitude}
\item{elev.in.mts}{Elevation in meters}
...
}
}
}
\source{
\url{https://www.ncei.noaa.gov/cdo-web/}
}
\usage{
precip_finalCA
}
\description{
Monthly precipitation for Sep - Dec, 2022
}
\keyword{datasets}
