% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt-gradients-est-plot.R
\name{spt_gradients_est_plot}
\alias{spt_gradients_est_plot}
\title{Plots estimated spatiotemporal differential processes}
\usage{
spt_gradients_est_plot(
  coords = NULL,
  data_frame = NULL,
  grid.points = NULL,
  gradients = NULL,
  col.y = NULL,
  col.grad = NULL,
  shape = NULL,
  point.size = 1,
  only.grad.no.curv = TRUE
)
}
\arguments{
\item{coords}{spatial coordinates}

\item{data_frame}{response values: supply only one time point at a time}

\item{grid.points}{grid points over which gradients/curvature is estimated}

\item{gradients}{data frame of estimated gradients}

\item{col.y}{significance color for response}

\item{col.grad}{significance color for gradients}

\item{shape}{should be NULL for now}

\item{point.size}{size of points to be plotted}

\item{only.grad.no.curv}{logical for only gradients}
}
\description{
Uses \code{ggplot2} to generate estimated plots for spatotemporal divergence and Laplacian
}
\keyword{spt_gradients_est_plot}
