% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt_gradients-sep.R
\name{spt_gradients_sep}
\alias{spt_gradients_sep}
\title{Posterior sampling of spatiotemporal differential processes}
\usage{
spt_gradients_sep(
  model = NULL,
  grid.points = NULL,
  cov.type.s = c("matern1", "matern2", "gaussian"),
  nbatch = NULL,
  plots = TRUE,
  only.grad.no.curv = FALSE,
  true = NULL
)
}
\arguments{
\item{model}{a list containing posterior samples from a collapsed or Gibbs sampler that fits a spatiotemporal Bayesian hierarchical model}

\item{grid.points}{can be an equi-spaced grid or in-fill or extrapolated points}

\item{cov.type.s}{three choices of covariance: Gaussian, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2})}

\item{nbatch}{batching for posterior samples. Defaults at 300.}

\item{plots}{logical for plotting. Defaults to FALSE.}

\item{only.grad.no.curv}{logical for only gradients and no curvature estimation}

\item{true}{true values, if available. Should be supplied as a list.}
}
\description{
Performs 1-for-1 sampling of spatiotemporal differential processes
}
\keyword{spt_gradients_sep}
