% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spt-plot.R
\name{spt_plot}
\alias{spt_plot}
\title{Plots estimated spatiotempoal differential processes}
\usage{
spt_plot(coords = NULL, Nt = NULL, data_frame = NULL, shape = NULL)
}
\arguments{
\item{coords}{spatial coordinates}

\item{Nt}{number of time points}

\item{data_frame}{response values: supply as a matrix of ncol = Nt}

\item{shape}{should be NULL for now}
}
\description{
Uses ggplot2 to generate estimated plots for spatotemporal divergence and Laplacian
}
\keyword{spt_plot}
