% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptwombling.R
\name{sptwombling}
\alias{sptwombling}
\title{Function for performing Bayesian spatiotemporal Wombling on surfaces}
\usage{
sptwombling(model = NULL, wombling.df = NULL)
}
\arguments{
\item{model}{a list containing posterior samples from a collapsed or Gibbs sampler that fits a spatiotemporal Bayesian hierarchical model}

\item{wombling.df}{a data-frame consisting of points, areas, normals for a triangulated surface}
}
\description{
Performs boundary analysis for surfaces. This uses a two-dimensional Riemann sum approximation.
}
\keyword{sptwombling}
