% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_cor_sepbl.R
\name{st_cor_sepbl}
\alias{st_cor_sepbl}
\title{Spatiotemporal Exponential Covariance Kernel}
\usage{
st_cor_sepbl(
  delta = NULL,
  Delta = NULL,
  phis = NULL,
  phit = NULL,
  cov.type.s = c("exponential", "gaussian", "matern1", "matern2"),
  cov.type.t = c("exponential", "gaussian", "matern1", "matern2", "poly")
)
}
\arguments{
\item{delta}{temporal distance matrix}

\item{Delta}{spatial distance matrix}

\item{phis}{spatial range}

\item{phit}{temporal range}

\item{cov.type.s}{spatial covariance type (four available choices: Exponential, Gaussian, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2})}

\item{cov.type.t}{spatial covariance type (five available choices: Exponential, Gaussian, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2}, Inverse/polynomial)}
}
\description{
Computes a separable exponential covariance kernel
}
\keyword{st_cor_sepbl}
