% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern.R
\name{st_cov_matern}
\alias{st_cov_matern}
\title{Spatiotemporal Mat\'ern Covariance Kernel with fractal parameter, \eqn{nu = 5/2}}
\usage{
st_cov_matern(
  delta = NULL,
  Delta = NULL,
  lphis = NULL,
  lphit = NULL,
  lsig2 = NULL,
  nu = NULL
)
}
\arguments{
\item{delta}{temporal distance matrix}

\item{Delta}{spatial distance matrix}

\item{lphis}{log spatial range}

\item{lphit}{log temporal range}

\item{lsig2}{log spatiotemporal variance parameter}

\item{nu}{fractal parameter}
}
\description{
Computes an adapted non-separable Mat\'ern covariance kernel with fractal parameter, \eqn{nu = 5/2}
}
\keyword{st_cov_matern}
