% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern3by2.R
\name{st_cov_matern1}
\alias{st_cov_matern1}
\title{Spatiotemporal Mat\'ern Covariance Kernel with fractal parameter, \eqn{nu = 3/2}}
\usage{
st_cov_matern1(
  delta = NULL,
  Delta = NULL,
  phis = NULL,
  phit = NULL,
  sig2 = NULL
)
}
\arguments{
\item{delta}{temporal distance matrix}

\item{Delta}{spatial distance matrix}

\item{phis}{spatial range}

\item{phit}{temporal range}

\item{sig2}{variance parameter}
}
\description{
Computes an adapted non-separable Mat\'ern covariance kernel with fractal parameter, \eqn{nu = 3/2}
}
\keyword{st_cov_matern1}
