% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_triangulate.R
\name{surf_triangulate}
\alias{surf_triangulate}
\title{Triangulates the wombling surface}
\usage{
surf_triangulate(
  curves_part = NULL,
  tr.points = 10,
  length.p = 10,
  plot.it = TRUE
)
}
\arguments{
\item{curves_part}{a list of partitioned curves}

\item{tr.points}{length of partition for triangular points. Defaults to 10}

\item{length.p}{length of partition for planar partitioned curves. Defaults to 10.}

\item{plot.it}{logical for whether to plot 3d points}
}
\description{
Triangulates the wombling surface
}
\keyword{surf_triangulate}
