% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcrossprod.R
\name{vcrossprod}
\alias{vcrossprod}
\title{Fucntion for vector cross-product}
\usage{
vcrossprod(u = NULL, v = NULL)
}
\arguments{
\item{u}{vector 1}

\item{v}{vector 2}
}
\description{
Computes a cross product for two vectors
}
\keyword{vcrossprod}
