% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlmBayes_spt.R
\name{hlmBayes_spt}
\alias{hlmBayes_spt}
\title{A spatiotemporal hierarchical Bayes Markov Chain Monte Carlo sampler}
\usage{
hlmBayes_spt(
  coords = NULL,
  t = NULL,
  y = NULL,
  X = NULL,
  D = NULL,
  z_init = NULL,
  phis_init = NULL,
  lower_phis = NULL,
  upper_phis = NULL,
  phit_init = NULL,
  lower_phit = NULL,
  upper_phit = NULL,
  sigma2_init = NULL,
  shape_sigma = NULL,
  scale_sigma = NULL,
  tau2_init = NULL,
  shape_tau = NULL,
  scale_tau = NULL,
  beta_init = NULL,
  mean_beta = NULL,
  prec_beta = NULL,
  steps_init = NULL,
  stept_init = NULL,
  niter = NULL,
  nburn = NULL,
  report = NULL,
  verbose = TRUE,
  cov.type = c("exponential", "gaussian", "matern1", "matern2")
)
}
\arguments{
\item{coords}{coordinates for observed process (order \eqn{N_s} x \eqn{2})}

\item{t}{temporal coordinates for observed process (order \eqn{N_t} x \eqn{1})}

\item{y}{observed response (order \eqn{L} x  \eqn{1})}

\item{X}{a matrix of covariates (order \eqn{L} x  \eqn{p})}

\item{D}{(use if replication at co-ordinate level) index for observations}

\item{z_init}{starting values of spatial effects (order \eqn{L} x  \eqn{1})}

\item{phis_init}{starting value for \eqn{\phi_s}}

\item{lower_phis}{lower bound for uniform prior on \eqn{\phi_s}}

\item{upper_phis}{upper bound for uniform prior on \eqn{\phi_s}}

\item{phit_init}{starting value for \eqn{\phi_t}}

\item{lower_phit}{lower bound for uniform prior on \eqn{\phi_t}}

\item{upper_phit}{upper bound for uniform prior on \eqn{\phi_t}}

\item{sigma2_init}{starting value for \eqn{\phi}}

\item{shape_sigma}{lower bound for uniform prior on \eqn{\phi}}

\item{scale_sigma}{upper bound for uniform prior on \eqn{\phi}}

\item{tau2_init}{starting value for \eqn{\phi}}

\item{shape_tau}{lower bound for uniform prior on \eqn{\phi}}

\item{scale_tau}{upper bound for uniform prior on \eqn{\phi}}

\item{beta_init}{starting value for \eqn{\phi}}

\item{mean_beta}{lower bound for uniform prior on \eqn{\phi}}

\item{prec_beta}{upper bound for uniform prior on \eqn{\phi}}

\item{steps_init}{tuning parameter for \eqn{\phi_s}}

\item{stept_init}{tuning parameter for \eqn{\phi_t}}

\item{niter}{number of MCMC iterations}

\item{nburn}{number of burn-in samples}

\item{report}{batch length}

\item{verbose}{if true prints output for batches}

\item{cov.type}{covariance type (three available choices: Gaussian, Mat\'ern(\eqn{\nu=3/2})), Mat\'ern(\eqn{\nu=5/2})}
}
\description{
Fits a univariate Gaussian spatiotemporal regression model: \eqn{Y(s,t)=x(s)^T\beta+Z(s)+\epsilon}. Parameters not listed are optional.
}
