% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{st_cov_exponential}
\alias{st_cov_exponential}
\title{Spatiotemporal Exponential Covariance Kernel}
\usage{
st_cov_exponential(
  delta = NULL,
  Delta = NULL,
  lphis = NULL,
  lphit = NULL,
  lsig2 = NULL
)
}
\arguments{
\item{delta}{temporal distance matrix}

\item{Delta}{spatial distance matrix}

\item{lphis}{log of spatial range}

\item{lphit}{log of temporal range}

\item{lsig2}{log of spatiotemporal variance parameter}
}
\description{
Computes an adapted non-separable exponential covariance kernel
}
\keyword{st_cov_gaussian}
